Events Api refactor for removing RegisteredEvents.java
