--- eventservice/src/main/java/com/socialservicefinder/eventservice/service/RegisteredEventsLookUpService.java
@@ -1,15 +1,0 @@
-package com.socialservicefinder.eventservice.service;
-
-import com.socialservicefinder.eventservice.dto.RegisteredEvents;
-
-import java.util.List;
-
-public interface RegisteredEventsLookUpService {
-    RegisteredEvents createRegisteredEvents(String userId);
-
-    RegisteredEvents save(RegisteredEvents registeredEvents);
-
-    List<String> fetchEventIdsByUserId(String userId);
-
-    void delete(RegisteredEvents registeredEvents);
-}
