+++ userservice/src/main/java/com/socialservicefinder/userservice/repository/UserRepository.java
@@ -4,6 +4,8 @@ import com.socialservicefinder.userservice.dto.User;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.mongodb.repository.Query;
 
-public interface UserRepository extends MongoRepository<User, Long>{
-    public User findUserByEmail(String email);
+public interface UserRepository extends MongoRepository<User, Long> {
+    User findUserByEmail(String email);
+
+    User findById(String userId);
 }
